video 04 (Theme Data)
In this video, it explains how we can use theme data to set text colors and use the scaffold's background property to change the background color. The copy With property is used to shorten the code length. Additionally, a dart file was created name input page, and the Stateful and Stateless classes were separated.