video 04(Themes)
In this video, it explains how we can use theme data to set text colors and use the scaffold's background property to change the background color. The copy With property is used to shorten the code length. Additionally, a dart file  create named  input page, and the Stateful and Stateless classes were separated.